requirements.txt
web3==6.17.0
pyyaml==6.0.1
pytest==8.2.0
requests==2.31.0
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
pandas==2.2.2
